+++ src/main/java/fasttui/component/LeafControl.java
@@ -38,6 +38,8 @@ public abstract class LeafControl extends Component implements Interactive {
 
     @Override
     public boolean contains(int cellX, int cellY) {
-        return cellX >= x && cellX < x + width && cellY >= y && cellY < y + height;
+        int absX = getAbsoluteX();
+        int absY = getAbsoluteY();
+        return cellX >= absX && cellX < absX + width && cellY >= absY && cellY < absY + height;
     }
 }